batch/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.12
langchain-core==0.2.29
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langsmith==0.1.98
multidict==6.0.5
numpy==1.26.4
openai==1.40.2
orjson==3.10.7
packaging==24.1
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.32
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
soupsieve==2.5
pytz==2024.1
beautifulsoup4==4.12.3
Werkzeug==3.0.3
MarkupSafe==2.1.5
Jinja2==3.1.4
itsdangerous==2.2.0
Flask-SQLAlchemy==3.1.1
Flask==3.0.3
Flask-Cors==4.0.1
click==8.1.7
blinker==1.8.2
gunicorn